roas-cli: split `preview` into its own module
Moves all preview-related production code (the `PreviewArgs` struct,
`REDOC_HTML` constant, `run_preview` entry point, `serve_preview_requests`
loop, and the small response helpers) and its three end-to-end / unit
tests out of `main.rs` and into a sibling `crates/roas-cli/src/preview.rs`.

`main.rs` now keeps only what it needs to wire the subcommand into clap:

  * `mod preview;` + `use preview::PreviewArgs;`
  * `Command::Preview(args) => preview::run_preview(args)` in the dispatch.
  * One `cli_parses_preview_subcommand` test confirming the parse goes
    through `Cli::try_parse_from` and yields the expected field values.

`read_and_parse` is promoted from private to `pub(crate)` so the preview
module can reuse the existing file-read + parse pipeline. `PreviewArgs`'s
fields are `pub(crate)` so the main-module clap-wiring test can still
assert on them.

No behaviour change — same 60 tests pass, dry-run publish stays clean.
